requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
librosa>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.18.0
matplotlib>=3.8.0
scipy>=1.11.0
soundfile>=0.12.0
python-dotenv>=1.0.0
